Move AssignLeaflets and AssignCurvedLeaflets to a new module 'leaflets'
