+++ src/lipyphilic/__init__.py
@@ -2,3 +2,8 @@ try:
     from lipyphilic._version import version as __version__
 except ImportError:
     __version__ = "not-installed"
+
+from lipyphilic.leaflets.assign_leaflets import (
+    AssignCurvedLeaflets,
+    AssignLeaflets,
+)
