+++ src/lipyphilic/lib/area_per_lipid.py
@@ -57,12 +57,12 @@ Example usage of :class:`AreaPerLipid`
 An MDAnalysis Universe must first be created before using AreaPerLipid::
 
   import MDAnalysis as mda
-  from lipyphilic.lib.assign_leaflets import AssignLeaflets
+  from lipyphilic.leaflets.assign_leaflets import AssignLeaflets
 
   u = mda.Universe(tpr, trajectory)
 
 Then we need to know which leaflet each lipid is in at each frame. This may be done using
-:class:`lipyphilic.lib.assign_leaflets.AssignLeaflets`::
+:class:`lipyphilic.leaflets.assign_leaflets.AssignLeaflets`::
 
   leaflets = AssignLeaflets(
     universe=u,
@@ -152,7 +152,7 @@ class AreaPerLipid(AnalysisBase):
         Tip
         ---
 
-        Leaflet membership can be determined using :class:`lipyphilic.lib.assign_leaflets.AssignLeaflets`.
+        Leaflet membership can be determined using :class:`lipyphilic.leaflets.assign_leaflets.AssignLeaflets`.
 
         """
         super().__init__(universe.trajectory)
